services/load-tests/bots/requirements.txt
kafka-python-ng>=2.2.2
prometheus-client>=0.21.0
